test: move GET /projects tests to separate file
